Unify block sanitization via shared BlockSanitizer class
Extracts sanitization logic into BlockSanitizer with recursive
innerHTML/innerContent/innerBlocks handling. Both EditPostBlocksAbility
and ReplacePostBlocksAbility now use the same sanitizeAndSerialize()
method instead of divergent approaches.

Fixes review feedback from PR #290.
